Pull request feedback, tidy image resizing
